Otimização de codigo
Deletei o sprite_classes.py e mudei a classe de criar quadrados para spritesheet.py

mudei mais algumas coisas no código spritesheet.py

comentei a mudança de posição do import spritesheet.py
