+++ spritesheets.py
@@ -1,6 +1,5 @@
 # Imports
 import pygame
-import spritesheet
 
 # Pygame setup
 pygame.init()
@@ -12,16 +11,19 @@ HEIGHT = 500
 screen = pygame.display.set_mode((WIDTH,HEIGHT))
 pygame.display.set_caption("Spritesheets")
 
-# Spritesheet
-sprite_sheet_image = pygame.image.load("assets/sprites/personagem/personagem_spritesheet.png").convert_alpha()
-sprite_sheet = spritesheet.SpriteSheet(sprite_sheet_image)
+# Spritesheet import
+'''  
+    sprite_sheet e sprite_sheet_image apenas funcionam depois da
+    definição do "pygame.display" e "pygame.init()"  
+'''
+import spritesheet
 
-BG = (50, 50, 50)
-BLACK = (0, 0, 0)
+BG = "#aaaaaa"
+BLACK = "#000000"
 
 
-frame_0 = sprite_sheet.get_image(0, 32,32, 3)
-frame_1 = sprite_sheet.get_image(1, 32,32, 3)
+frame_0 = spritesheet.sprite_sheet.get_image(0, 32,32, 3)
+frame_1 = spritesheet.sprite_sheet.get_image(1, 32,32, 3)
 
 # Framerate
 clock = pygame.time.Clock()
